bump next to the newest canary version
