+++ apps/earn-protocol-institutions/package.json
@@ -4,8 +4,8 @@
   "license": "Apache-2.0",
   "scripts": {
     "prebuild": "concurrently \"pnpm:codegen\" \"pnpm:generate-module-types\"",
-    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3004 --turbopack\" \"pnpm:generate-module-types-watch\"",
-    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build",
+    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3004\" \"pnpm:generate-module-types-watch\"",
+    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build --webpack",
     "start": "dotenv -e ../../.env -e ../../.env.local -- next start -p 3004",
     "lint": "tsc --noEmit && eslint .",
     "lint:fix": "eslint . --fix",
@@ -49,13 +49,13 @@
     "graphql-request": "^6.1.0",
     "jose": "^6.0.12",
     "lodash-es": "^4.17.1",
-    "next": "15.5.10",
+    "next": "16.2.0-canary.77",
     "pg": "^8.14.1",
     "qrcode": "1.5.4",
     "react-day-picker": "9.9.0",
-    "react-dom": "19.1.0",
+    "react-dom": "19.2.3",
     "react-toastify": "^11.0.3",
-    "react": "19.1.0",
+    "react": "19.2.3",
     "recharts": "3.7.0",
     "sharp": "^0.33.5",
     "usehooks-ts": "^3.1.0",
@@ -73,8 +73,8 @@
     "@types/node": "^20.14.2",
     "@types/pg": "^8.11.11",
     "@types/qrcode": "1.5.5",
-    "@types/react-dom": "19.0.3",
-    "@types/react": "19.0.8",
+    "@types/react-dom": "19.2.2",
+    "@types/react": "19.2.3",
     "graphql": "^16.8.1",
     "knip": "5.64.0",
     "postcss": "^8",
