+++ apps/earn-protocol-landing-page/package.json
@@ -4,8 +4,8 @@
   "license": "Apache-2.0",
   "scripts": {
     "prebuild": "pnpm generate-module-types",
-    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3003 --turbopack\" \"pnpm:generate-module-types-watch\"",
-    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build",
+    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3003\" \"pnpm:generate-module-types-watch\"",
+    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build --webpack",
     "start": "next start -p 3003",
     "lint": "eslint .",
     "knip": "knip",
@@ -25,10 +25,10 @@
     "concurrently": "^9.0.1",
     "lodash-es": "^4.17.1",
     "mixpanel-browser": "^2.53.0",
-    "next": "15.5.10",
+    "next": "16.2.0-canary.77",
     "react-animate-height": "^3.2.3",
-    "react-dom": "19.1.0",
-    "react": "19.1.0",
+    "react-dom": "19.2.3",
+    "react": "19.2.3",
     "sharp": "^0.33.5",
     "zod": "3.25.61"
   },
@@ -39,8 +39,8 @@
     "@types/lodash-es": "^4.17.1",
     "@types/mixpanel-browser": "^2.49.0",
     "@types/node": "^20.14.2",
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3",
+    "@types/react": "19.2.3",
+    "@types/react-dom": "19.2.2",
     "knip": "5.64.0",
     "typed-css-modules": "0.9.1",
     "typescript": "^5.7.3"
