+++ apps/earn-protocol/package.json
@@ -4,8 +4,8 @@
   "license": "Apache-2.0",
   "scripts": {
     "prebuild": "concurrently \"pnpm:codegen\" \"pnpm:generate-module-types\"",
-    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3002 --turbopack\" \"pnpm:generate-module-types-watch\"",
-    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build",
+    "dev": "dotenv -e ../../.env -e ../../.env.local -- concurrently \"next dev -p 3002\" \"pnpm:generate-module-types-watch\"",
+    "build:apps": "dotenv -e ../../.env -e ../../.env.local -- next build --webpack",
     "start": "dotenv -e ../../.env -e ../../.env.local -- next start -p 3002",
     "lint": "tsc --noEmit && eslint .",
     "lint:fix": "eslint . --fix",
@@ -53,13 +53,13 @@
     "lodash-es": "^4.17.1",
     "mixpanel-browser": "^2.53.0",
     "mixpanel": "^0.18.0",
-    "next": "15.5.10",
+    "next": "16.2.0-canary.77",
     "pg": "^8.14.1",
     "react-animate-height": "^3.2.3",
-    "react-dom": "19.1.0",
+    "react-dom": "19.2.3",
     "react-infinite-scroller": "^1.2.6",
     "react-toastify": "^11.0.3",
-    "react": "19.1.0",
+    "react": "19.2.3",
     "recharts": "3.7.0",
     "sharp": "^0.33.5",
     "viem": "2.43.5",
@@ -74,8 +74,8 @@
     "@types/lodash-es": "^4.17.1",
     "@types/mixpanel-browser": "^2.49.0",
     "@types/node": "^20.14.2",
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3",
+    "@types/react": "19.2.3",
+    "@types/react-dom": "19.2.2",
     "@types/react-infinite-scroller": "^1.2.5",
     "graphql": "^16.8.1",
     "knip": "5.64.0",
