+++ packages/app-earn-ui/package.json
@@ -30,8 +30,8 @@
     "@types/loadable__component": "^5.13.9",
     "@types/lodash-es": "^4.17.1",
     "@types/node": "20.12.7",
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3",
+    "@types/react": "19.2.3",
+    "@types/react-dom": "19.2.3",
     "@vitejs/plugin-react-swc": "^3.8.0",
     "concurrently": "^9.0.1",
     "eslint": "8.57.0",
@@ -59,10 +59,10 @@
     "embla-carousel-react": "8.3.0",
     "embla-carousel": "8.3.0",
     "lodash-es": "^4.17.1",
-    "next": "15.5.10",
+    "next": "16.2.0-canary.77",
     "react-animate-height": "^3.2.3",
-    "react-dom": "19.1.0",
-    "react": "19.1.0",
+    "react-dom": "19.2.3",
+    "react": "19.2.3",
     "react-day-picker": "9.9.0",
     "react-toastify": "^11.0.3",
     "uniqolor": "1.1.1",
