+++ packages/app-icons/package.json
@@ -37,7 +37,7 @@
     "@summerfi/app-types": "workspace:*",
     "@summerfi/eslint-config": "workspace:*",
     "@summerfi/typescript-config": "workspace:*",
-    "@types/react": "19.0.8",
+    "@types/react": "19.2.3",
     "eslint": "8.57.0",
     "glob": "^11.0.0",
     "jest": "29.7.0",
@@ -46,9 +46,9 @@
   },
   "peerDependencies": {
     "@loadable/component": "^5.16.4",
-    "next": "15.5.10",
-    "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "next": "16.2.0-canary.77",
+    "react": "19.2.3",
+    "react-dom": "19.2.3"
   },
   "publishConfig": {
     "access": "public"
