+++ packages/app-risk/package.json
@@ -20,7 +20,7 @@
     "@summerfi/eslint-config": "workspace:*",
     "@summerfi/typescript-config": "workspace:*",
     "@types/node": "^22.14.1",
-    "@types/react": "19.0.8",
+    "@types/react": "19.2.3",
     "@vitejs/plugin-react-swc": "^3.8.0",
     "eslint": "8.57.0",
     "glob": "^11.0.0",
@@ -32,8 +32,8 @@
   "peerDependencies": {
     "@summerfi/app-utils": "workspace:*",
     "kysely": "^0.27.3",
-    "next": "15.5.10",
-    "react": "19.1.0",
+    "next": "16.2.0-canary.77",
+    "react": "19.2.3",
     "zod": "3.25.61"
   },
   "publishConfig": {
