+++ packages/app-utils/package.json
@@ -49,7 +49,7 @@
     "dayjs": "^1.11.12",
     "jose": "^6.0.10",
     "lodash-es": "^4.17.1",
-    "next": "15.5.10",
+    "next": "16.2.0-canary.77",
     "viem": "2.43.5"
   },
   "dependencies": {
