+++ sdk/sdk-client-react/package.json
@@ -31,6 +31,6 @@
     "@types/react": "^19.0.8"
   },
   "peerDependencies": {
-    "react": "19.1.0"
+    "react": "19.2.3"
   }
 }
